Renamed resources to routes
