+++ backend/src/main/kotlin/routes/RoutesUtils.kt
@@ -1,4 +1,4 @@
-package res
+package routes
 
 import io.ktor.application.ApplicationCall
 import io.ktor.http.HttpStatusCode
@@ -6,7 +6,7 @@ import io.ktor.request.header
 import io.ktor.response.respond
 import services.UsersService
 
-object ResourcesUtil {
+object RoutesUtils {
 
     private val usersService = UsersService()
 
@@ -19,12 +19,6 @@ object ResourcesUtil {
         val uid = usersService.getUserId(token)
         if (uid == null) call.respond(HttpStatusCode.Unauthorized)
         else block(uid)
-    }
-
-    suspend fun withTaskId(call: ApplicationCall, block: suspend (taskId: Int) -> Unit) {
-        val taskId = call.parameters["id"]?.toInt()
-        if (taskId == null) call.respond(HttpStatusCode.BadRequest)
-        else block(taskId)
     }
 
 }
